requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#

django-filter==2.0.0
django==2.1.3
djangorestframework==3.9.0
pytz==2018.7
